refactor(game): extract per-frame run loop to game/main-loop.ts (R2.5)
Move the run loop out of snowglider.ts into `src/game/main-loop.ts` as
`createMainLoop(deps)`, returning `updateSnowman` / `updateCamera` / `animate` /
`startLoop` / `handleResize`. The loop now owns its private `lastTime`; the three
lifecycle sites that previously did `lastTime = performance.now(); animate(lastTime)`
call the new `startLoop()` instead (same seed → first delta ~0). The window-resize
handler moves with the loop and is re-registered by the coordinator.

`updateSnowman` / `updateCamera` are still re-published on `window` by
`publishGameGlobals()` (the coordinator destructures them from the loop). The initial
one-shot `cameraManager.initialize(...)` stays in the coordinator.

Also removes the dead, write-only `lastTechnique` variable (declared + assigned in
`resetSnowman` and `updateSnowman`, never read anywhere in src/ or tests) — keeping
it would have meant injecting a no-op setter across the new module boundary.

Mechanical, behavior-preserving: per-frame ordering is unchanged. Dropped the now-
unused coordinator imports `updateStatsHud`, `ShakeOffset`, and
`AVALANCHE_TRIGGER_DISTANCE`.

Verification: lint (0 errors) / typecheck / test (contract guard 44/44) / build
green; puppeteer 87/87 across 7/7 suites (gameplay/camera/avalanche/regression drive
the loop).
